renamed to ingredientModels
