+++ backend/src/Sauces.Api/Models/Requests/FermentationRecipeRequest.cs
@@ -1,11 +1,11 @@
 using System.Text.Json.Serialization;
-using Sauces.Api.Models;
-using Sauces.Core.Model;
+
+namespace Sauces.Api.Models.Requests;
 
 public class FermentationRecipeRequest
 {
     [JsonPropertyName("ingredients")]
-    public List<IngredientsModel> Ingredients { get; set; }
+    public List<IngredientModel> Ingredients { get; set; }
     
     [JsonPropertyName("lengthInDays")]
     public int LengthInDays { get; set; }
